Test sensor values
In this step, I want to verify the correct operation of all the connected sensors.
This is why I also changed the name of this source file to "testvalues" instead of "testDHT22" as it was named before.
The first task is to properly define each pin on the ESP32 for the respective sensor.