Organizing and sensors file
In this step, I have created a file that will handle all sensor values, so if any adjustments are needed in the future, it won't affect the main file. This also includes creating 'sensor.h' to follow best practices.
I have also implemented functions in the main file to read values from the sensors as soon as the device is connected to a network.
If sensor values are received, it indicates a successful connection.
Additionally, I have implemented error handling at each stage to make it easy to troubleshoot if anything goes wrong.