Connect to AWS IoT Core
I have adjusted to ensure that we obtain accurate readings from all sensors and have been working on calibrating the water level sensor.
Afterward, I have been working on setting up the appropriate functions and configurations to connect the ESP32 to AWS IoT Core.
This has involved creating a 'thing' in AWS and generating certificates with suitable policies to ensure that the connection is always secure and that all values are not at risk of manipulation.
The next step is to create functions that can send values in the most appropriate manner according to best practices.